rename error wrapper file
